LayoutTests/fast/animation/request-animation-frame-throttling-lowPowerMode-expected.txt

Test that requestAnimationFrame gets throttled in low power mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS framesPerSecond < 35 is true
PASS successfullyParsed is true

TEST COMPLETE
